+++ app/controllers/slice_controller.rb
@@ -35,7 +35,9 @@ class SliceController < ApplicationController
     @subject = current_user&.subjects&.find_by(project: @project)
     pdf_file = Rails.root.join(@project.generate_printed_pdf!(@subject))
     if File.exist?(pdf_file)
-      send_file(pdf_file, filename: "#{@project.name.titleize.gsub(/\s/, "")}ConsentForm.pdf", type: "application/pdf", disposition: "inline")
+      File.open(pdf_file, "r") do |f|
+        send_data f.read, filename: "#{@project.name.titleize.gsub(/\s/, "")}ConsentForm.pdf", type: "application/pdf", disposition: "inline"
+      end
     else
       redirect_to slice_consent_path(@project), alert: "Unable to generate PDF at this time."
     end
@@ -64,7 +66,9 @@ class SliceController < ApplicationController
     @data = @subject.data(variables)
     pdf_file = Rails.root.join(@subject.generate_overview_report_pdf!(@data))
     if File.exist?(pdf_file)
-      send_file(pdf_file, filename: "#{@project.name.titleize.gsub(/\s/, "")}SleepReport.pdf", type: "application/pdf", disposition: "inline")
+      File.open(pdf_file, "r") do |f|
+        send_data f.read, filename: "#{@project.name.titleize.gsub(/\s/, "")}SleepReport.pdf", type: "application/pdf", disposition: "inline"
+      end
     else
       redirect_to slice_overview_report_path(@project), alert: "Unable to generate PDF at this time."
     end